requirements.txt
alembic==1.12.0
bcrypt==4.0.1
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
Deprecated==1.2.14
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
flask-msearch==0.2.9.2
Flask-Reuploaded==1.3.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
gcloud==0.18.3
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==2.0.2
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.5.1
Mako==1.2.4
MarkupSafe==2.1.3
msgpack==1.0.7
oauth2client==4.1.3
packaging==23.2
pdfkit==1.0.0
proto-plus==1.23.0
protobuf==4.25.1
psycopg2==2.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.1
PyJWT==2.8.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-jwt==4.1.0
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.20
SQLAlchemy-Utils==0.41.1
stripe==6.5.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==2.3.7
wkhtmltopdf==0.2
wrapt==1.16.0
WTForms==3.0.1
